requirements.txt
# dashboard
streamlit>=1.30
streamlit-autorefresh==1.0.1

# data
pandas>=2.0
numpy>=1.24

# postgres connectivity (pick one; psycopg is modern)
psycopg[binary]>=3.1

# synthetic CSV generator
Faker>=20.0
